requirements.txt
django
django-parler
django-ckeditor
Pillow
